move functions from pkg/util/helper/work.go to pkg/util/work.go
